disallowed worker to reach certain pages

+++ keliuPrieziura/src/Controller/CipherController.php
@@ -9,17 +9,29 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use App\Service\AdminService;
 
 /**
  * @Route("/cipher")
  */
 class CipherController extends AbstractController
 {
+    private $adminService;
+    public function __construct(AdminService $adminService)
+    {
+        $this->adminService = $adminService;
+    }
     /**
      * @Route("/", name="cipher_index", methods={"GET"})
      */
     public function index(CipherRepository $cipherRepository): Response
     {
+        if (!$this->adminService->isAdmin()) {
+
+            $response =  $this->redirect('/', 301);
+            $response->setCache(['max_age' => 0]);
+            return $response;
+        }
         return $this->render('cipher/index.html.twig', [
             'ciphers' => $cipherRepository->findAll(),
         ]);
@@ -30,6 +42,12 @@ class CipherController extends AbstractController
      */
     public function new(Request $request): Response
     {
+        if (!$this->adminService->isAdmin()) {
+
+            $response =  $this->redirect('/', 301);
+            $response->setCache(['max_age' => 0]);
+            return $response;
+        }
         $cipher = new Cipher();
         $form = $this->createForm(CipherType::class, $cipher);
         $form->handleRequest($request);
@@ -53,6 +71,11 @@ class CipherController extends AbstractController
      */
     public function show(Cipher $cipher): Response
     {
+        if (!$this->adminService->isAdmin()) {
+            $response =  $this->redirect('/', 301);
+            $response->setCache(['max_age' => 0]);
+            return $response;
+        }
         return $this->render('cipher/show.html.twig', [
             'cipher' => $cipher,
         ]);
@@ -63,6 +86,11 @@ class CipherController extends AbstractController
      */
     public function edit(Request $request, Cipher $cipher): Response
     {
+        if (!$this->adminService->isAdmin()) {
+            $response =  $this->redirect('/', 301);
+            $response->setCache(['max_age' => 0]);
+            return $response;
+        }
         $form = $this->createForm(CipherType::class, $cipher);
         $form->handleRequest($request);
 
@@ -83,7 +111,12 @@ class CipherController extends AbstractController
      */
     public function delete(Request $request, Cipher $cipher): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$cipher->getId(), $request->request->get('_token'))) {
+        if (!$this->adminService->isAdmin()) {
+            $response =  $this->redirect('/', 301);
+            $response->setCache(['max_age' => 0]);
+            return $response;
+        }
+        if ($this->isCsrfTokenValid('delete' . $cipher->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($cipher);
             $entityManager->flush();
